SBC/README: Fix backwards description of the VP GND jumper
Issue #29 from ea4eoz noted that this is backwards for the revision H
board. It's clearly must be backwards for _every_ revision, since
closing a jumper, rather than opening it, is the way to connect the
other side of the jumper to ground.

(This error was inserted by me when I originally wrote this section
of the documentation.)